requirements.txt
bleak
mediapipe
pygame
opencv-python
requests
python-dotenv
numpy
